feat: promote the wasm_next api to default
This patch removes the old wasm api and promotes the new one.
