+++ src/lib.rs
@@ -11,7 +11,5 @@ mod shared;
 mod stdlib;
 mod visitors;
 mod wasm;
-#[cfg(feature = "wasm_next")]
-mod wasm2;
 
 pub use server::LspServer;
